refactor: badge component
