+++ src/components/ui/Badge/Badge.tsx
@@ -1,30 +1,8 @@
-import React from 'react';
-import { cva, VariantProps } from 'class-variance-authority';
-
+import { BadgeProps } from './types';
 import { cn } from '../../../utils/cn';
+import { badgeVariants } from './constant/Badge';
 
-export const badgeVariants = cva(
-  'rounded leading-none inline-flex items-center justify-center capitalize py-1 px-2 text-xs gap-1.5',
-  {
-    variants: {
-      variant: {
-        gray: 'bg-gray-100 text-gray-800',
-        yellow: 'bg-yellow-100 text-yellow-800',
-        red: 'bg-red-100 text-red-800',
-        green: 'bg-green-100 text-green-800',
-        blue: 'bg-blue-100 text-blue-800',
-      },
-    },
-    defaultVariants: {
-      variant: 'gray',
-    },
-  },
-);
-export interface BadgeProps
-  extends React.HTMLAttributes<HTMLSpanElement>,
-    VariantProps<typeof badgeVariants> {}
-
-export const Badge: React.FC<BadgeProps> = ({
+export const Badge = ({
   className,
   variant,
   children,
