+++ src/components/ui/Badge/types/index.ts
@@ -1,0 +1,6 @@
+import { VariantProps } from 'class-variance-authority';
+import { badgeVariants } from '../constant/Badge';
+
+export interface BadgeProps
+  extends React.HTMLAttributes<HTMLSpanElement>,
+    VariantProps<typeof badgeVariants> {}
